requirements.txt
argon2-cffi==23.1.0
email-validator==2.2.0
fastapi==0.115.4
minio==7.2.10
passlib[argon2]==1.7.4
pillow==10.4.0
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
redis==5.1.1
reportlab==4.2.5
SQLAlchemy==2.0.36
uvicorn[standard]==0.32.0
pytest==8.3.3
pytest-asyncio==0.24.0
